refactor(license): split getCoreGraceExpiresAt into its own file
